Move api, cli and config tests to src
