+++ packages/backend/.mocharc.json
@@ -1,8 +1,8 @@
 {
-  "spec": "test/**/*.test.ts",
+  "spec": "{src,test}/**/*.test.ts",
   "require": "esbuild-register",
   "watchExtensions": "ts",
   "extension": "ts",
-  "file": "test/setup.ts",
+  "file": "src/test/setup.ts",
   "reporterOption": "maxDiffSize=0"
 }
