+++ packages/backend/src/api/controllers/tvl/generateTvlApiResponse.test.ts
@@ -9,10 +9,10 @@ import {
 } from '@l2beat/types'
 import { expect } from 'earljs'
 
-import { asNumber } from '../../../../src/api/controllers/tvl/asNumber'
-import { generateTvlApiResponse } from '../../../../src/api/controllers/tvl/generateTvlApiResponse'
-import { AggregateReportRecord } from '../../../../src/peripherals/database/AggregateReportRepository'
-import { ReportRecord } from '../../../../src/peripherals/database/ReportRepository'
+import { AggregateReportRecord } from '../../../peripherals/database/AggregateReportRepository'
+import { ReportRecord } from '../../../peripherals/database/ReportRepository'
+import { asNumber } from './asNumber'
+import { generateTvlApiResponse } from './generateTvlApiResponse'
 
 describe(generateTvlApiResponse.name, () => {
   it('returns the correct groupings', () => {
